backend/src/auth/requirements.txt
fastapi[all]
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-consul
prometheus-fastapi-instrumentator
python-dotenv